requirements.txt
conformer==0.3.2
Cython==3.0.2
diffusers==0.21.2
espnet==202304
fastapi==0.95.1
ffmpy==0.3.0
Flask==2.3.2
flask_restful==0.3.9
g2p_en==2.1.0
librosa==0.9.2
loguru==0.7.2
matplotlib==3.6.0
num2words==0.5.12
numba==0.57.0
numpy==1.23.4
pickle5==0.0.11
py_vncorenlp==0.1.4
pydub==0.25.1
PyYAML==6.0
requests==2.28.1
scikit_learn==1.2.2
scipy==1.10.1
tensorboard
tgt==1.4.4
tqdm==4.63.0
Unidecode==1.3.4
uvicorn==0.22.0
waitress==2.1.2